review comments addressed, deleted obsolete DummyTest, moved TestProject into its own file
