--- plugins/com.google.cloud.tools.eclipse.appengine.libraries.test/src/com/google/cloud/tools/eclipse/appengine/libraries/DummyTest.java
@@ -1,11 +1,0 @@
-package com.google.cloud.tools.eclipse.appengine.libraries;
-
-import org.junit.Test;
-
-public class DummyTest {
-
-  @Test
-  public void test() {
-    // make surefire happy
-  }
-}
